Osobny plik i filmy od razu
Dodany plik library.js oraz wyświetlanie od razu filmów po wejściu na library
